shanyraq_p2_backend_project_3/pyproject.toml
[project]
name = "shanyrak-p1-backend-project-2"
version = "0.1.0"
description = ""
authors = [
    {name = "molspace",email = "nurlybek.amangeldiev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
